Testing the HID
Something seems to be missing and the WM_INPUT case never happens in the
MSG_PROC no matter what.

However the Observer pattern stuff seems to work fine. Tried to send a
message over the links hardcoded and it worked.

All that is needed is to figure out why WM_INPUT is "broken".